Change shader_creates file length to a long int.
This was why comments weren't being accepted in lane.fs, as the file was being cut off and causing syntax errors.

+++ src/shader.c
@@ -58,7 +58,7 @@ void shader_create(GLuint *shader, GLenum type, const char *source_path)
     assert(source_file);
 
     // the length of the file
-    uint8_t length;
+    long int length;
 
     // the buffer that the file is read into
     char *buffer;
